refactor: consolidate tRPC client configuration to eliminate duplication
- Extract shared configuration utilities to packages/trpc/react/config.ts
- Update packages/trpc/react/trpc.ts to use shared config
- Update apps/web/app/_trpc/trpc-client.ts to use shared config
- Update apps/web/app/_trpc/query-client.ts to use shared config
- Export new utilities from packages/trpc/react/index.ts

This eliminates the need to update both places when adding new tRPC routers or changing routing logic.
