+++ packages/trpc/react/index.ts
@@ -3,3 +3,5 @@ export type { RouterInputs, RouterOutputs } from "./trpc";
 export { trpc, transformer } from "./trpc";
 
 export { ENDPOINTS } from "./shared";
+
+export { getTrpcUrl, createTrpcLinks, createQueryClientConfig, resolveEndpoint } from "./config";
